ai-service/pyproject.toml
[project]
name = "ghostypedia-ai-service"
version = "0.1.0"
description = "AI service layer for Ghostypedia using Google Gemini"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.3.2",
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
